feat: add sanity verification script and update data preparation to use single-position holdout for detection experiments.

+++ generalization-detect/data/detect/prepare.py
@@ -31,7 +31,7 @@ LENGTH = 20            # fixed input length (number of digit slots before ':')
 #   'single'   : hold out one position (HELDOUT_SINGLE)
 #   'even_odd' : train on even positions, test on odd positions
 #   'half'     : train on first half, test on second half
-SPLIT = 'half'
+SPLIT = 'single'
 HELDOUT_SINGLE = 12    # only used when SPLIT == 'single'
 
 N_TRAIN = 50_000       # number of training examples
